API refactoring
This commit refactors the metadata tags to be present under the `Tags`
property of the `MP4File` object. This will make for a more consistent API
for other file data.

Additionally, this commit changes the semantics of the `MP4File` object,
introducing a static `Open()` method and making the constructor private
rather than using a constructor to specify the file, and a comprehensive
`Save()` method to write all changes to a file rather than having to
manage edits of each part of the file information piecemeal. We also
introduce a `Load()` method, which will refresh the file information. I'm
not sure if the `Load()` method will remain a public part of the API, so
don't rely on it.