Adding tracking of editing of chapter properties
In keeping with only writing changes to a file when the file changes, we
need to track when the properties of a chapter change. This will mark the
ChapterList as dirty only when a property is added, deleted, or modified,
whether that modification is done in-place, or out of band.